Move class 'RemoteActionListeners' to messages package
Aim is for better cohesion as the listeners are closely related
to the message types.
