+++ http-clients/src/main/java/org/triplea/http/client/remote/actions/messages/server/ServerRemoteActionMessageType.java
@@ -5,7 +5,6 @@ import static org.triplea.http.client.web.socket.messages.MessageTypeListenerBin
 import java.net.InetAddress;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import org.triplea.http.client.remote.actions.RemoteActionListeners;
 import org.triplea.http.client.web.socket.messages.MessageTypeListenerBinding;
 import org.triplea.http.client.web.socket.messages.WebsocketMessageType;
