+++ http-clients/src/test/java/org/triplea/http/client/remote/actions/messages/server/ServerRemoteActionMessageTypeTest.java
@@ -10,7 +10,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.triplea.http.client.IpAddressParser;
-import org.triplea.http.client.remote.actions.RemoteActionListeners;
 
 @ExtendWith(MockitoExtension.class)
 class ServerRemoteActionMessageTypeTest {
